% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpbal.R
\name{gpbal}
\alias{gpbal}
\title{Compute optimally balanced Gaussian process propensity scores}
\usage{
gpbal(X, y, cov_function, init_theta = NULL, verbose = F,
  balance_metric = "mom_sq", approx_method = "ep",
  ep_vers = "parallel", estimand = "ATE", eta = 0.05, tol = 0.01)
}
\arguments{
\item{X}{Matrix of covariates to be included in the analysis and balanced on}

\item{y}{Set of observed treatment assignments (y in (0,1))}

\item{cov_function}{Covariance matrix; for examples, see \code{sqexp} or similar}

\item{verbose}{Decision to print progress to screen - default TRUE}

\item{ep_vers}{Sequential or Parallel EP Algorthim - default \code{parallel}, alternative \code{sequential}}

\item{tol}{Tolerance of algorithms.  Difference between the latent scores at each iteration - default 1e-2}

\item{max_iters}{Maximum number of iterations of algorithm - default 20}
}
\value{
Object that contains the weights obtained from the balancing procedure and parameters from the optimization procedure

The object that is returned is a list that contains the following entries
\itemize{
\item{ \code{Number_Iters} - Number of iterations for algorithm}
\item{ \code{PosteriorMean} - Posterior mean of latent scores}
\item{ \code{PosteriorVar} - Posterior covariance of latent scores}
\item{ \code{tilde_nu} - }
\item{ \code{tilde_tau} - }
\item{ \code{log_Z_ep} - EP Approximation to Log Likelihood}
\item{ \code{ComputationTime} - Runtime of EP algorithm for fixed covariance matrix}
\item{ \code{thetas} - optimal theta from optimization routine}
\item{ \code{ps} - Probit transformed posterior mean}
}
}
\description{
Compute optimally balanced Gaussian process propensity scores
}
\examples{
n_obs <- 500
X1 <- rnorm(n_obs)
X2 <- rnorm(n_obs)
p <- pnorm( 0.5 * X1 + 0.5 * X2 )
TA <- rbinom(n_obs, 1, p)
dat <- data.frame(X1 = X1, X2 = X2, TA = TA)
covmat <- sqexp(cbind(X1, X2))
system.time(res <- gpbal_fixed(TA, covmat))
plot(res$ps, p, pch = 19, col = rgb(0,0,0,0.5))

}
